+++ internal/bridge/chain/zano/deposit.go
@@ -33,7 +33,7 @@ func (p *Client) GetDepositData(id db.DepositIdentifier) (*db.DepositData, error
 		return nil, bridgeTypes.ErrDepositNotFound
 	}
 
-	if len(transaction.ServiceEntries) < id.TxNonce+1 {
+	if int64(len(transaction.ServiceEntries)) < id.TxNonce+1 {
 		return nil, bridgeTypes.ErrDepositNotFound
 	}
 	addr, chainId, err := parseDestinationData(transaction.ServiceEntries[id.TxNonce])
